Add a facility to send ICMP errors from arbitrary points in the stack.
Replace the udp ICMP sending code with the generic facility.
Also IPv6 wants to sent ICMP packets while processing extension headers
but until now there was no easy way to do so. That will come in
a following commit.

There are several places where we need to find out what kind of
transport protocol header is in the header storage and we don't want to
have to parse the IPV6 extension headers again just to find out. Generally
speaking we are not yet using it on output but make sure it is valid as
we will probably use it in the filtering.

Updates #2211.

Startblock:
   any of
      has LGTM from ghanan
      has LGTM from bhaskerh
   and then
   involve tamird
PiperOrigin-RevId: 325959491